added pages and updated content
in the about page i updated the content and the slides is working well. I added an instituite page and placed buttons for the institutes, i still have to arrange styling. I also added a course list page.
note: still need to work the navigation and have to update and add new colors to the page